Refactoring generateFactoryClass
First pass at refactoring `generateFactoryClass` to make it work with KSP and Embedded.
Currently I am using a single test, so it is possible that this iteration does not yet fulfill all requirements for all the tests.
`gFC2` is the WIP, `gFC` is just a reference that will be deleted.